Remove deprecated client, suspend compression, and dead code; update buffer to 3.0.0
- Delete old DefaultWebSocketClient and Frame data class
- Rename ModularWebSocketClient → DefaultWebSocketClient
- Remove suspend compression/decompression functions (~250 lines)
- Remove suspend close helpers (closeIfNeeded/closeAll)
- Simplify Factory: remove WebSocketImplementation enum
- Remove dead code: unused ParsedFrame factory, CloseCode constants,
  WebSocketClient convenience methods, Opcode.isValid()
- Delete CompressionTest and FrameSerializationTest (belong in buffer module)
- Update buffer/buffer-compression to 3.0.0
