+++ gradle/libs.versions.toml
@@ -7,8 +7,8 @@ mavenPublish = "0.35.0"
 kotlinxCoroutines = "1.10.2"
 kotlinWrappers = "2025.12.6"
 kotlinxBenchmark = "0.4.15"
-buffer = "2.3.1-SNAPSHOT"
-bufferCompression = "2.3.1-SNAPSHOT"
+buffer = "3.0.0"
+bufferCompression = "3.0.0"
 socket = "1.2.6-SNAPSHOT"
 androidxCore = "1.17.0"
 androidxTestRunner = "1.7.0"
