+++ src/commonMain/kotlin/com/ditchoom/websocket/WebSocketClient.kt
@@ -6,7 +6,6 @@ import com.ditchoom.buffer.SuspendCloseable
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
 
 interface WebSocketClient : SuspendCloseable {
     val scope: CoroutineScope
@@ -24,16 +23,6 @@ interface WebSocketClient : SuspendCloseable {
     suspend fun ping(payloadData: ReadBuffer = EMPTY_BUFFER)
 
     suspend fun isPingSupported(): Boolean = true
-
-    fun onIncomingWebsocketMessage(cb: (WebSocketMessage) -> Unit) =
-        scope.launch {
-            incomingMessages.collect { cb(it) }
-        }
-
-    fun onConnectionStateChange(cb: (ConnectionState) -> Unit) =
-        scope.launch {
-            connectionState.collect { cb(it) }
-        }
 
     companion object
