+++ src/commonMain/kotlin/com/ditchoom/websocket/frame/FrameReader.kt
@@ -511,117 +511,6 @@ sealed interface ParsedFrame {
 }
 
 /**
- * Factory function to create ParsedFrame instances (for backwards compatibility).
- * Prefer using the concrete types directly when possible.
- */
-@Suppress("FunctionName")
-fun ParsedFrame(
-    fin: Boolean,
-    rsv1: Boolean,
-    rsv2: Boolean,
-    rsv3: Boolean,
-    opcode: Opcode,
-    masked: Boolean,
-    payloadLength: Int,
-    payload: ReadBuffer,
-): ParsedFrame {
-    // Construct header bytes from individual flags
-    var byte1 = opcode.value.toInt() and 0x0F
-    if (fin) byte1 = byte1 or 0x80
-    if (rsv1) byte1 = byte1 or 0x40
-    if (rsv2) byte1 = byte1 or 0x20
-    if (rsv3) byte1 = byte1 or 0x10
-    val header1 = FrameHeaderByte1(byte1)
-
-    var byte2 = payloadLength.coerceAtMost(127) and 0x7F
-    if (masked) byte2 = byte2 or 0x80
-    val header2 = FrameHeaderByte2(byte2)
-
-    return when (opcode) {
-        Opcode.Text -> ParsedFrame.DataFrame.Text(header1, header2, payloadLength, payload)
-        Opcode.Binary -> ParsedFrame.DataFrame.Binary(header1, header2, payloadLength, payload)
-        Opcode.Continuation -> ParsedFrame.DataFrame.Continuation(header1, header2, payloadLength, payload)
-        Opcode.Ping -> ParsedFrame.ControlFrame.Ping(header1, header2, payloadLength, payload)
-        Opcode.Pong -> ParsedFrame.ControlFrame.Pong(header1, header2, payloadLength, payload)
-        Opcode.Close -> {
-            // Parse close code and reason from payload if present
-            val result = parseClosePayload(payload, payloadLength)
-            ParsedFrame.ControlFrame.Close(
-                header1,
-                header2,
-                payloadLength,
-                payload,
-                result.code,
-                result.reason,
-                result.hasInvalidUtf8,
-            )
-        }
-        else -> {
-            // Reserved opcodes (0x3-0x7 non-control, 0xB-0xF control) are invalid per RFC 6455
-            ParsedFrame.InvalidFrame(
-                header1,
-                header2,
-                payloadLength,
-                payload,
-                "Reserved opcode $opcode is not allowed",
-            )
-        }
-    }
-}
-
-/**
- * Result of parsing a close frame payload.
- */
-private data class ClosePayloadResult(
-    val code: CloseCode,
-    val reason: String,
-    val hasInvalidUtf8: Boolean,
-)
-
-/**
- * Parses close code and reason from a close frame payload.
- */
-private fun parseClosePayload(
-    payload: ReadBuffer,
-    payloadLength: Int,
-): ClosePayloadResult {
-    if (payloadLength == 0) {
-        return ClosePayloadResult(CloseCode.NO_STATUS_RECEIVED, "", false)
-    }
-
-    // Save position for restoration
-    val savedPos = payload.position()
-
-    if (payloadLength == 1) {
-        // Invalid: close frame with only 1 byte
-        payload.position(savedPos)
-        return ClosePayloadResult(CloseCode.PROTOCOL_ERROR, "Invalid close payload length", false)
-    }
-
-    val code = CloseCode(payload.readUnsignedShort())
-    var hasInvalidUtf8 = false
-    val reason =
-        if (payload.hasRemaining()) {
-            try {
-                payload.readString(payload.remaining(), Charset.UTF8)
-            } catch (
-                @Suppress("TooGenericExceptionCaught") e: Throwable,
-            ) {
-                // Use Throwable to catch JS TypeError from TextDecoder
-                hasInvalidUtf8 = true
-                ""
-            }
-        } else {
-            ""
-        }
-
-    // Restore position for consumers who want to read the raw payload
-    payload.position(savedPos)
-
-    return ClosePayloadResult(code, reason, hasInvalidUtf8)
-}
-
-/**
  * Exception thrown when frame parsing fails.
  */
 class FrameParseException(
