+++ src/commonMain/kotlin/com/ditchoom/websocket/frame/FrameWriter.kt
@@ -393,4 +393,3 @@ class FrameWriter(
         }
     }
 }
-
