+++ src/jsMain/kotlin/com/ditchoom/websocket/WebSocket.kt
@@ -12,14 +12,4 @@ actual fun WebSocketClient.Companion.allocate(
     pool: BufferPool,
     parentScope: CoroutineScope?,
     allocationZone: AllocationZone,
-    implementation: WebSocketImplementation,
-): WebSocketClient =
-    when (implementation) {
-        WebSocketImplementation.DEFAULT ->
-            try {
-                DefaultWebSocketClient(connectionOptions, pool, parentScope, allocationZone)
-            } catch (e: UnsupportedOperationException) {
-                BrowserWebSocketController(connectionOptions, pool, parentScope, allocationZone)
-            }
-        WebSocketImplementation.MODULAR -> ModularWebSocketClient(connectionOptions, pool, parentScope, allocationZone)
-    }
+): WebSocketClient = DefaultWebSocketClient(connectionOptions, pool, parentScope, allocationZone)
